aligned new and edit book screens
